pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openweights"
version = "0.5.0"
description = "An openai-like sdk for finetuning and batch inference"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "supabase",
    "python-dotenv",
    "httpx>=0.24.0",
    "fire",
    "runpod",
    "PyJWT",
    "fastapi",
    "uvicorn",
    "openai",
    "huggingface-hub",
    "diskcache",
    "scp",
    "httpx[http2]",
    "ipython>=9.3.0",
    "ipykernel>=6.29.5",
    "jupyterlab>=4.4.3",
    "cachier",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black",
    "isort",
    "pre-commit",
]
worker = [
    "torch>=2.7.1",
    "transformers>=4.52.4",
    "unsloth>=2024.8",
    "tokenizers>=0.21.1",
    "datasets>=3.6.0",
    "trl>=0.18.1",
]
[tool.hatch.build.targets.wheel]
packages = ["openweights"]
